requirements.txt
cassandra-driver
commlib-py
confluent-kafka==2.3.0
docker==6.1.3
docker-compose==1.29.2
fastapi==0.68.2
kafka-python==2.0.2
numpy
paho-mqtt==1.6.1
pydantic==1.10.4
requests==2.31.0
uvicorn==0.15.0
websocket-client==1.6.1
